requirements.txt
Flask==1.1.1
pymongo==3.10.1
gunicorn==20.0.4
python-dotenv==0.12.0